+++ middle/middle-ram.cpp
