+++ middle/middle.cpp
